LPS-161717 Take into account if there is a layout copy in process in order to restore the previous subscriptions. If the portlet is removed, then the old subscriptions are removed.

+++ modules/apps/subscription/subscription-service/src/main/java/com/liferay/subscription/internal/model/listener/SubscriptionPortletPreferencesModelListener.java
@@ -14,11 +14,18 @@
 
 package com.liferay.subscription.internal.model.listener;
 
+import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+import com.liferay.portal.kernel.dao.orm.RestrictionsFactoryUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.ModelListener;
 import com.liferay.portal.kernel.model.PortletPreferences;
+import com.liferay.portal.kernel.service.ClassNameLocalService;
+import com.liferay.portal.kernel.service.PortletPreferencesLocalService;
+import com.liferay.portal.kernel.transaction.TransactionCommitCallbackUtil;
+import com.liferay.portal.kernel.util.CopyLayoutThreadLocal;
+import com.liferay.subscription.model.Subscription;
 import com.liferay.subscription.service.SubscriptionLocalService;
 
 import org.osgi.service.component.annotations.Component;
@@ -42,10 +49,80 @@ public class SubscriptionPortletPreferencesModelListener
 		}
 
 		try {
-			_subscriptionLocalService.deleteSubscriptions(
-				portletPreferences.getCompanyId(),
-				portletPreferences.getModelClassName(),
-				portletPreferences.getPortletPreferencesId());
+			if (!CopyLayoutThreadLocal.isCopyLayout()) {
+				_subscriptionLocalService.deleteSubscriptions(
+					portletPreferences.getCompanyId(),
+					portletPreferences.getModelClassName(),
+					portletPreferences.getPortletPreferencesId());
+			}
+			else {
+				TransactionCommitCallbackUtil.registerCallback(
+					() -> {
+						PortletPreferences remainingPortletPreferences =
+							_portletPreferencesLocalService.
+								fetchPortletPreferences(
+									portletPreferences.getOwnerId(),
+									portletPreferences.getOwnerType(),
+									portletPreferences.getPlid(),
+									portletPreferences.getPortletId());
+
+						if (remainingPortletPreferences == null) {
+							_subscriptionLocalService.deleteSubscriptions(
+								portletPreferences.getCompanyId(),
+								portletPreferences.getModelClassName(),
+								portletPreferences.getPortletPreferencesId());
+						}
+						else {
+							ActionableDynamicQuery actionableDynamicQuery =
+								_subscriptionLocalService.
+									getActionableDynamicQuery();
+
+							actionableDynamicQuery.setAddCriteriaMethod(
+								dynamicQuery -> {
+									dynamicQuery.add(
+										RestrictionsFactoryUtil.eq(
+											"companyId",
+											remainingPortletPreferences.
+												getCompanyId()));
+
+									dynamicQuery.add(
+										RestrictionsFactoryUtil.eq(
+											"classNameId",
+											_classNameLocalService.
+												getClassNameId(
+													remainingPortletPreferences.
+														getModelClassName())));
+
+									dynamicQuery.add(
+										RestrictionsFactoryUtil.eq(
+											"classPK",
+											portletPreferences.
+												getPortletPreferencesId()));
+								});
+
+							actionableDynamicQuery.setPerformActionMethod(
+								(Subscription subscription) -> {
+									subscription.setClassPK(
+										remainingPortletPreferences.
+											getPortletPreferencesId());
+
+									_subscriptionLocalService.
+										updateSubscription(subscription);
+								});
+
+							try {
+								actionableDynamicQuery.performActions();
+							}
+							catch (Exception exception) {
+								_log.error(
+									"Unable to restore subscriptions",
+									exception);
+							}
+						}
+
+						return null;
+					});
+			}
 		}
 		catch (Exception exception) {
 			_log.error("Unable to delete subscriptions", exception);
@@ -54,6 +131,12 @@ public class SubscriptionPortletPreferencesModelListener
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		SubscriptionPortletPreferencesModelListener.class);
+
+	@Reference
+	private ClassNameLocalService _classNameLocalService;
+
+	@Reference
+	private PortletPreferencesLocalService _portletPreferencesLocalService;
 
 	@Reference
 	private SubscriptionLocalService _subscriptionLocalService;
